Update analysis.R
In this update, my goal is to input all the scripts of the movies into the data frame. In order to achieve this I did two things:
Between the 17th to the 19th lines, I'm Importing one srt file.
Between the 27th and the 32 line I'm creating a for loop, in order to automate the importing and cleaning process of the str files.